refactor: move the master token secret into server state
`FromRequest::Config` uses the `Default` trait to create dependencies,
which is no good for us because we don't want to use a default value for
the master token secret. Instead, this change pushes the master secret
into the `ServerState` struct, where we can set it after reading from
environment variables or a config file.

One slight downside to this is that I had to wrap it in an `Arc`, to
avoid cloning a fresh copy of the master secret in every instance of
`ServerState`. I did try giving `ServerState` an explicit lifetime and
borrowing the master secret as a slice, but I couldn't get that to play
nicely with the requirement for a static lifetime on the `App` instance.