feat: load default and env-specific settings from the file system
Pulls out the default settings to `config/default.json` and also tries
to load environment-specific settings from `config/$SYNC_ENV.json`. My
thinking is that this will be useful when we want to point at a standard
GCP region/zone for dev or CI environments.

There's also an attempt to load from `config/local.json`, which is added
to `.gitignore` so that there is the option to locally set master_secret
or a password for the MySQL connection string without accidentally
committing them to the repo.

Lastly there is a quick check added to make sure `master_secret` is set
to something in prod.