Fix purge_ttl timeout
Sometimes, the default purge_ttl takes so long it times out and gets killed.

This patch adds SYNC_INCREMENTAL=1 to purge_ttl. When this is passed, a different method for deleting is employed where 1000 expired records are fetched, then explicitly deleted by id. This uses much smaller transactions and can slowly catch up on a backlog of expired records.

Fix #570