Add libmysqlclient dependency solution for Debian (#847)
On Debian, there is not libmysqlclient available since Stretch. Debian has fully migrated to MariaDB as replacement. There is a libmariadb-dev package, which ships libmariadbclient.so, but this is currently not accepted by the migrations_macros build. A simple solution is the available compat package: https://packages.debian.org/libmariadb-dev-compat

Signed-off-by: MichaIng <micha@dietpi.com>