f pause: rework pagination for Sort::Index
Index is provided by the client and is independent of the modification or creation time. This means that the pagination should use that value rather than the time.

Offset has been reworked to specify and "index"(either modification or sortindex) and "offset" (bid), but some testing is still failing (notably mysql get_bso_aaa) which is failing to get the correct BSOs for a pagination of "2:b3" (it's failing to return the right listings and count (order by off?))

Need to also check spanner to make sure that this is reflected correctly into that code base.

TODO:
    * Fix pagination for sort::Index
    * remove COPIOUS debugging code.
    * verify e2e tests still work.