feat: postgres db pool and session implementation (#1875)
We need to create the Postgres specific Db struct and associated methods. The majority of API methods will be implemented upon this struct, so this is to initialize the main logic so that the traits can be implemented.

This also encompasses additions to the Postgres-specific database pool implementation.

Several of the trait bounds (methods) have to be satisfied, but this stubs them out so that the PgDb has a common interface that is the same as those for MySQL and Spanner.

Doc comments added in shared error.rs crate for clarity, since it is a shared interface for all databases.

Removed dead code annotations for clippy that are no longer needed in Tokenserver.

Closes STOR-399